+++ internal/features/periods/blocks/form_templ.go
@@ -78,6 +78,10 @@ func PeriodForm(view dto.PeriodFormView, schedules []scheduleDTO.PersonWithSched
 		if templ_7745c5c3_Err != nil {
 			return templ_7745c5c3_Err
 		}
+		templ_7745c5c3_Err = forms.DeleteButton("period", view.ID).Render(ctx, templ_7745c5c3_Buffer)
+		if templ_7745c5c3_Err != nil {
+			return templ_7745c5c3_Err
+		}
 		templ_7745c5c3_Err = forms.CancelButton("/periods").Render(ctx, templ_7745c5c3_Buffer)
 		if templ_7745c5c3_Err != nil {
 			return templ_7745c5c3_Err